Add Docker and a PostgreSQL database service
Set up Docker and a PostgreSQL database service that can be run via
`docker compose`.

The database service may be configured via environment variables and
with volumes (such as a 'postgresql.conf' file).

This creates a development environment that makes it easy to interact
with a real database connection and persist data during development as
needed.

The `justfile` received a couple of convenience commands that spawn
interactive `psql` or `bash` sessions.