Make Logger more flexible and add dependencies The global Logger instance is initialized by the function `setup_logger` inside of 'src/logger.rs'. It was renamed from `init` because it now takes arguments in the form of environment variables.
This change helps make the logger more flexible as it can be
instantiated differently depending on the type of environment that it is
running in. The old function used purely hard-coded values.

The `main` function was modified to reflect the changes made to the
logger. The values that are passed to the logger now reside here.

Dependencies required for database interaction were added:
    - `sqlx`: Allows interacting with Postgresql (supports async)
    - `uuid`: Generate UUIDs that can be used for database model id's