Renamed CircleCI job and changed filter values The 'build' job was renamed to 'build-and-test' to more accurately represent what the job is actually doing. In addition to this, the alias names for some of the 'security' job steps were changed to be more concise.
The 'workflows' top-level entry has been modified such that the CI jobs
will only be executed by CircleCI when commits are made to the `main`
and/or the `release` branches. This change is temporary.

As the project grows, the changes made to the filters will be reverted
and the CI workflow will execute on every commit.